Window 3
Added the ability for a window to minimized, maximized, focused, full screened and unfullscreened in the code.

Added an override so that minimized windows have their main loops skipped.

Added a monitor class which wraps a GLFWmonitor. Some of the information in this new monitor class is needed for fullscreening a window